Add mechanism for passing global configurations
The biggest change in this commit is the addition of `src/config.py`, which acts as a
singleton module to pass global configs around SPRAS. When the Snakefile is run, it
uses it's magically available `config` to instantiate the config object in the config
module. After instantiation, other modules can `import src.config as config`, which allows
them to access various config parameters. For example, to access the container framework,
you can do:
```
import src.config as config
framework = config.config.framework
```

In addition to modularizing configuration, I also added the ability to specify container
registries and owners in the config. A full container name looks like
`docker.io/reedcompbio/pathlinker:latest`. The config option considers docker.io the base_url,
reedcompbio the owner, and then pathlinker will be hardcoded by that method. Lacking in this
implementation is the ability to specify exactly what tag to run each PRM against. With the new
config module, that should be relatively simple to add.

As far as implentation choices, since the config module makes various values globally available,
I decided to stop passing some values like `singularity` down through nested levels of function
calls, and instead get those values close to where they're needed.
