+++ spras/pathlinker.py
@@ -8,7 +8,9 @@ from spras.interactome import (
     reinsert_direction_col_directed,
 )
 from spras.prm import PRM
-from spras.util import prepare_volume, run_container
+from spras.util import prepare_volume
+from spras.containers import run_container
+import spras.config as config
 
 __all__ = ['PathLinker']
 
@@ -67,7 +69,7 @@ class PathLinker(PRM):
 
     # Skips parameter validation step
     @staticmethod
-    def run(nodetypes=None, network=None, output_file=None, k=None, singularity=False):
+    def run(nodetypes=None, network=None, output_file=None, k=None):
         """
         Run PathLinker with Docker
         @param nodetypes:  input node types with sources and targets (required)
@@ -115,10 +117,10 @@ class PathLinker(PRM):
 
         print('Running PathLinker with arguments: {}'.format(' '.join(command)), flush=True)
 
-        # TODO consider making this a string in the config file instead of a Boolean
-        container_framework = 'singularity' if singularity else 'docker'
+        container_framework = config.config.framework
+        container_suffix = "pathlinker"
         out = run_container(container_framework,
-                            'reedcompbio/pathlinker',
+                            container_suffix,
                             command,
                             volumes,
                             work_dir)
