+++ test/LocalNeighborhood/test_ln.py
@@ -3,6 +3,9 @@ from filecmp import cmp
 from pathlib import Path
 
 import pytest
+import spras.config as config
+
+config.init_from_file("config/config.yaml")
 
 # TODO consider refactoring to simplify the import
 # Modify the path because of the - in the directory
