+++ test/analysis/test_summary.py
@@ -3,6 +3,9 @@ from pathlib import Path
 import pandas as pd
 
 from spras.analysis.summary import summarize_networks
+import spras.config as config
+
+config.init_from_file("config/config.yaml")
 
 # Notes:
 # - Column labels are required in the node table
